(BSR)[PRO] fix: fix circular dependency
