+++ pro/src/pages/AdageIframe/app/components/AdageDiscovery/types.ts
@@ -1,0 +1,8 @@
+import { AdagePlaylistType } from "apiClient/adage"
+
+export type TrackerElementArg = {
+  playlistId: number
+  playlistType: AdagePlaylistType
+  elementId?: number
+  index?: number
+}
